refactor(BA-6616): move initiate_vfolder_deletion to the repository layer
Relocate the procedural vfolder deletion routine out of
models/vfolder/row.py into repositories/vfolder/deletion.py, removing the
models -> repositories reverse dependency (the lazy purger import). The
function is now imported directly by the group/user db_sources.
