+++ src/ai/backend/manager/models/vfolder/__init__.py
@@ -25,7 +25,6 @@ from .row import (
     get_allowed_vfolder_hosts_by_group,
     get_allowed_vfolder_hosts_by_user,
     get_sessions_by_mounted_folder,
-    initiate_vfolder_deletion,
     is_mount_duplicate,
     is_unmanaged,
     prepare_vfolder_mounts,
@@ -68,7 +67,6 @@ __all__ = (
     "get_allowed_vfolder_hosts_by_group",
     "get_allowed_vfolder_hosts_by_user",
     "get_sessions_by_mounted_folder",
-    "initiate_vfolder_deletion",
     "is_mount_duplicate",
     "is_unmanaged",
     "prepare_vfolder_mounts",
