+++ src/ai/backend/manager/models/vfolder/row.py
@@ -19,7 +19,6 @@ from typing import (
     override,
 )
 
-import aiotools
 import sqlalchemy as sa
 import trafaret as t
 from sqlalchemy.dialects import postgresql as pgsql
@@ -64,7 +63,6 @@ from ai.backend.manager.errors.api import InvalidAPIParameters
 from ai.backend.manager.errors.common import ObjectNotFound
 from ai.backend.manager.errors.storage import (
     InsufficientStoragePermission,
-    VFolderGone,
     VFolderNotFound,
     VFolderOperationFailed,
     VFolderPermissionError,
@@ -135,7 +133,6 @@ __all__: Sequence[str] = (
     "filter_host_allowed_permission",
     "get_allowed_vfolder_hosts_by_group",
     "get_allowed_vfolder_hosts_by_user",
-    "initiate_vfolder_deletion",
     "prepare_vfolder_mounts",
     "query_accessible_vfolders",
     "update_vfolder_status",
@@ -1352,72 +1349,6 @@ async def delete_vfolder_relation_rows(
         await _delete_vfolder_permission_rows(db_session, vfolder_row_ids)
 
     await execute_with_txn_retry(_delete, begin_session, db_conn)
-
-
-async def initiate_vfolder_deletion(
-    db_engine: ExtendedAsyncSAEngine,
-    requested_vfolders: Sequence[VFolderDeletionInfo],
-    storage_manager: StorageSessionManager,
-    _storage_ptask_group: aiotools.PersistentTaskGroup | None = None,
-    *,
-    force: bool = False,
-) -> int:
-    """Purges VFolder content from storage host."""
-    # Lazy import to avoid circular import
-    from ai.backend.manager.repositories.base.purger import BatchPurger, execute_batch_purger
-    from ai.backend.manager.repositories.vfolder.purgers import (
-        VFolderInvitationBatchPurgerSpec,
-        VFolderPermissionBatchPurgerSpec,
-    )
-
-    vfolder_info_len = len(requested_vfolders)
-    vfolder_ids = tuple(vf_id.folder_id for vf_id, _, _ in requested_vfolders)
-    if vfolder_info_len == 0:
-        return 0
-
-    async with db_engine.begin_session() as db_session:
-        await execute_batch_purger(
-            db_session,
-            BatchPurger(spec=VFolderInvitationBatchPurgerSpec(vfolder_ids=vfolder_ids)),
-        )
-        await execute_batch_purger(
-            db_session,
-            BatchPurger(spec=VFolderPermissionBatchPurgerSpec(vfolder_ids=vfolder_ids)),
-        )
-    await update_vfolder_status(
-        db_engine,
-        vfolder_ids,
-        VFolderOperationStatus.DELETE_ONGOING,
-        do_log=False,
-        force=force,
-    )
-
-    already_deleted: list[VFolderDeletionInfo] = []
-
-    for vfolder_info in requested_vfolders:
-        folder_id, host_name, unmanaged_path = vfolder_info
-        proxy_name, volume_name = storage_manager.get_proxy_and_volume(
-            host_name, is_unmanaged(unmanaged_path)
-        )
-        try:
-            manager_client = storage_manager.get_manager_facing_client(proxy_name)
-            await manager_client.delete_folder(volume_name, str(folder_id))
-        except (VFolderOperationFailed, InvalidAPIParameters) as e:
-            if e.status == 410:
-                already_deleted.append(vfolder_info)
-        except VFolderGone:
-            already_deleted.append(vfolder_info)
-    if already_deleted:
-        vfolder_ids = tuple(vf_id.folder_id for vf_id, _, _ in already_deleted)
-
-        await update_vfolder_status(
-            db_engine, vfolder_ids, VFolderOperationStatus.DELETE_COMPLETE, do_log=False
-        )
-        log.info("vfolders already deleted {}", [str(x) for x in vfolder_ids])
-
-    log.info("Started purging vfolders {}", [str(x) for x in vfolder_ids])
-
-    return vfolder_info_len
 
 
 async def ensure_quota_scope_accessible_by_user(
