+++ src/ai/backend/manager/repositories/group/db_source/db_source.py
@@ -66,7 +66,6 @@ from ai.backend.manager.models.vfolder import (
     VFolderDeletionInfo,
     VFolderRow,
     VFolderStatusSet,
-    initiate_vfolder_deletion,
     vfolder_status_map,
     vfolders,
 )
@@ -109,6 +108,7 @@ from ai.backend.manager.repositories.group.types import (
 )
 from ai.backend.manager.repositories.permission_controller.creators import UserRoleCreatorSpec
 from ai.backend.manager.repositories.permission_controller.role_manager import RoleManager
+from ai.backend.manager.repositories.vfolder.deletion import initiate_vfolder_deletion
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
